+++ src/fields/fp.rs
@@ -11,7 +11,7 @@ use lazy_static::lazy_static;
 #[cfg(feature = "bits")]
 use ff::{FieldBits, PrimeFieldBits};
 
-use crate::arithmetic::{SqrtTableHelpers, adc, mac, sbb};
+use crate::arithmetic::{SqrtTableHelpers, VartimeField, adc, mac, sbb};
 #[cfg(feature = "deferred")]
 use crate::deferred::{DeferredField, Product};
 
@@ -606,6 +606,12 @@ impl ff::Field for Fp {
     }
 }
 
+impl VartimeField for Fp {
+    fn invert_vartime(&self) -> Option<Self> {
+        super::modinv62::invert::<super::modinv62::FpParams>(&self.0).map(Self)
+    }
+}
+
 impl ff::PrimeField for Fp {
     type Repr = [u8; 32];
 
@@ -928,11 +934,16 @@ fn test_root_of_unity() {
 #[test]
 fn test_inv_root_of_unity() {
     assert_eq!(Fp::ROOT_OF_UNITY_INV, Fp::ROOT_OF_UNITY.invert().unwrap());
+    assert_eq!(
+        Fp::ROOT_OF_UNITY_INV,
+        Fp::ROOT_OF_UNITY.invert_vartime().unwrap()
+    );
 }
 
 #[test]
 fn test_inv_2() {
     assert_eq!(Fp::TWO_INV, Fp::from(2).invert().unwrap());
+    assert_eq!(Fp::TWO_INV, Fp::from(2).invert_vartime().unwrap());
 }
 
 #[test]